// src/interfaces/film.interface.ts
export interface Film {
  
      title: String;
      episode_id: String;
      opening_crawl: String;
      director: String;
      producer: String;
      release_date: Date;
      characters: Array<String>;
      planets: Array<String>;
      starships: Array<String>;
      vehicles: Array<String>;
      species: Array<String>;
  
  }
  