// src/interfaces/people.interface.ts
export interface People {
  name: String;
  height: Number;
  mass: Number;
  hair_color: String;
  skin_color: String;
  eye_color: String;
  birth_year: String;
  gender: String;
  homeworld: String;
  films: Array<String>;
  species: Array<String>;
  vehicles: Array<String>;
  starships: Array<String>;
  id: String;
}